+++ rmr/models/fields/hash_lookup.py
@@ -1,8 +1,6 @@
-import crcmod
-
 from django.db import models
 
-make_crc = crcmod.predefined.mkPredefinedCrcFun('crc-64')
+from rmr.utils.hash import crc64
 
 
 class HashLookup(models.BigIntegerField):
@@ -11,12 +9,6 @@ class HashLookup(models.BigIntegerField):
         self.object_field = object_field
         super().__init__(*args, **kwargs)
 
-    def make_hash(self, obj):
-        crc = make_crc(bytes(obj, encoding='utf-8'))
-        if crc > self.MAX_BIGINT:
-            crc = -(crc >> 1)
-        return crc
-
     def deconstruct(self):
         name, path, args, kwargs = super().deconstruct()
         kwargs['object_field'] = self.object_field
@@ -24,7 +16,7 @@ class HashLookup(models.BigIntegerField):
 
     def pre_save(self, model_instance, add):
         obj = getattr(model_instance, self.object_field)
-        return None if obj is None else self.make_hash(obj)
+        return None if obj is None else crc64(obj)
 
     def get_prep_lookup(self, lookup_type, value):
-        return self.make_hash(value)
+        return crc64(value)