+++ rmr/utils/iterate.py
@@ -1,5 +1,5 @@
+import collections
 import itertools
-from collections import deque
 
 
 def split_every(iterable, chunk_size):
@@ -10,4 +10,4 @@ def split_every(iterable, chunk_size):
         first_item = next(chunk)  # raises StopIteration when `iterator` is empty
         chunk = itertools.chain((first_item,), chunk)
         yield chunk
-        deque(chunk, 0)  # Exhaust unused `chunk`
+        collections.deque(chunk, 0)  # Exhaust unused `chunk`